Completed 29. First Navigator - Appswitch Navigator
This one gave me incredible grief. Had to add
react-navigator and it did not come in cleanly.
Also had problems with the build process going in
a loop. Found out this had something to do with files
residing in dropbox. Moved the project out of dropbox and
problem solved, but not have no backup for this project
or any other project using react-native. Will look for
a better solution.